heads

+++ src/circle.cpp
@@ -2,7 +2,7 @@
 #include <iostream>;
 using namespace std;
 
-long long float pi = 3,1415926;
+float pi = 3,1415926;
 bool flag = 0;
 
 circle::circle(float r){
